AuxiliaryFunctions

Documentation

ranksDescending :: [Rank]

Purpose:

The Ranks in Descending order, Haskell lets me do [Two .. Ace] but not [Ace .. Two]

So this is a workaround because I dont want to do reverse [Two .. Ace]

Parameters: N/A

Output:

Returns a List containing the Ranks in Descending order

compareCardsByRankAsc :: Card -> Card -> Ordering

Purpose:

Compares two cards for sorting in ascending order, with Ace being considered the highest rank

Parameters:

1st Parameter (x) : The first card we are comparing with the second card

2nd Parameter (y) : The second card we are comparing with the first card

Output:

Returns an Ordering which will result in a list of cards being sorted in ascending order by rank if used with sortBy

Notes:

This function is intended to be used by the sortBy function

compareCardsByRankAscAceLow :: Card -> Card -> Ordering

Purpose:

Compares two cards for sorting in ascending order, with Ace being considered the lowest rank

Parameters:

1st Parameter (x) : The first card we are comparing with the second card

2nd Parameter (y) : The second card we are comparing with the first card

Output:

Returns an Ordering which will result in a list of cards being sorted in ascending order by rank if used with sortBy

Notes:

This function is intended to be used by the sortBy function

compareCardsByRankDesc :: Card -> Card -> Ordering

Purpose:

Compares two cards for sorting in descending order, with Ace being considered the highest rank

Parameters:

1st Parameter (x) : The first card we are comparing with the second card

2nd Parameter (y) : The second card we are comparing with the first card

Output:

Returns an Ordering which will result in a list of cards being sorted in ascending order by rank if used with sortBy

Notes:

This function is intended to be used by the sortBy function

findCardsWithRank :: Rank -> Hand -> Hand

Purpose:

Finds all the Cards of a certain rank from a list of cards

Parameters:

1st Parameter (target_rank) : The rank that we are searching for

2nd Parameter (hand) : The list of cards to search from

Output:

Returns a list of cards containing only cards with the same rank as "target_rank"

findCardsWithSuit :: Suit -> Hand -> Hand

Purpose:

Finds all the Cards of a certain suit from a list of cards

Parameters:

1st Parameter (target_suit) : The suit that we are searching for

2nd Parameter (hand) : The list of cards to search from

Output:

Returns a list of cards containing only cards with the same suit as "target_suit"

isHandAscending :: Hand -> Bool

Purpose:

Checks if a list of cards only contains elements which have successive ranks to each other

Parameters:

1st Parameter : The list of cards we are checking against

Output:

Returns true if the list only contains elements which have successive ranks to each other

Returns false otherwise

Notes:

The input list should be sorted before being input into this function

Helper function for contains when we are checking if the hand type is a Straight

getScoringCards :: Hand -> HandType -> [Card]

Purpose:

Gets a list of cards which are contributing to the score for a given hand type

Parameters:

1st Parameter (xs) : The list of cards we are checking against

2nd Parameter : The hand type we are checking the scoring cards for

Output:

Returns a list of cards which are contributing to the score for a given hand type

Notes:

Only Hands with 5 cards or less should be input to this function

Helper function for whichCardsScore

combinations :: Int -> [a] -> [[a]]

Purpose:

Generates a list of all possible combinations of size n from a list

Parameters:

1st Parameter (n) : size of combinations

2nd Parameter (xs) : The list to generate the combinations from

Output:

Returns a list of all possible combinations of size n from a list

Notes:

Helper function for highestScoringHand

cardsFromMove :: Move -> [Card]

Purpose:

Extracts the list of cards that were played/discarded in the move

Parameters:

1st Parameter : The Move to extract the cards from

Output:

Returns the list of cards that were played/discarded in the move

Notes:

Helper function for probabilityBasedAI

moveType :: Move -> PlayOrDiscard

Purpose:

Extracts whether we played or discarded in a move

Parameters:

1st Parameter : The Move to extract the information from

Output:

Returns whether we played or discarded in a move

Notes:

Helper function for probabilityBasedAI