CourseworkOne

Documentation

contains :: Hand -> HandType -> Bool

Purpose:

Check whether a played hand is a certain hand type

Parameters:

1st Parameter (xs) : The list of cards to check from

2nd Parameter : Hand type to check against

Output:

Returns whether a played hand is a certain hand type

Note:

Only Hands with 5 cards or less should be input to this function

bestHandType :: Hand -> HandType

Purpose:

Identify the highest value hand type in a played hand

Parameters:

1st Parameter (xs) : The list of cards to check from

Output:

Returns the highest value hand type in a played hand

whichCardsScore :: Hand -> [Card]

Purpose:

Identify cards in hand the that contribute to the best hand type

Parameters:

1st Parameter (xs) : The list of cards to check from

Output:

Returns a list containing the cards in hand the that contribute to the best hand type

scoreHand :: Hand -> Int

Purpose:

Calculates the score of a hand based on the formula [(base score + sum of rank scores) * multiplier]

Parameters:

1st Parameter (xs) : The list of cards to check from

Output:

Returns the score of a hand

highestScoringHand :: [Card] -> Hand

Purpose:

Identify the highest scoring hand of 5 cards out of n>=5 cards by scoring all combinations of size 5 and playing the highest scoring combination

Parameters:

1st Parameter (xs) : The list of cards to check from

Output:

Returns the highest scoring hand of 5 cards out of n>=5 cards

simpleAI :: [Move] -> [Card] -> Move

Purpose:

An AI implementation for Halatro where we always play the 5 highest ranked cards

Parameters:

1st Parameter (moves) : Previous Moves that have been played

2nd Parameter (cards) : The cards in our current hand

Output:

Returns the move we should play

sensibleAI :: [Move] -> [Card] -> Move

Purpose:

An AI implementation for Halatro where we always play the best hand

Parameters:

1st Parameter (moves) : Previous Moves that have been played

2nd Parameter (cards) : The cards in our current hand

Output:

Returns the move we should play

probabilityBasedAI :: Double -> [Move] -> [Card] -> Move

Purpose:

An AI implementation for Halatro where we discard if our probability of improving the current hand is greater than a threshold value

When we play, we play the cards returned by highestScoringHand

When we discard, we discard as many cards as possibe given that the rank of the card is less than a Ten

Parameters:

1st Parameter (pMin) : Discard probability threshold

2nd Parameter (moves) : Previous Moves that have been played

3rd Parameter (cards) : The cards in our current hand

Output:

Returns the move we should play

Notes:

See probabilityOfImprovement and probabilityOfHand for more information on probability calculations

Helper function for myAI

myAI :: [Move] -> [Card] -> Move

Purpose:

My Halatro AI implementation, uses probabilityBasedAI

Parameters:

1st Parameter (moves) : Previous Moves that have been played

2nd Parameter (cards) : The cards in our current hand

Output:

Returns the move we should play