Halatro.Types

Documentation

data Rank

The rank of a card, from Two to Ace

Constructors

Two

Three

Four

Five

Six

Seven

Eight

Nine

Ten

Jack

Queen

King

Ace

shortRank :: Rank -> String

A shorter version of the rank, for use in the Terminal UI.

data Suit

The suit of a card

Constructors

Clubs

Diamonds

Hearts

Spades

shortSuit :: Suit -> String

This will only render properly for fonts that include the Unicode card suits, which *should* be most of them, but Windows command prompt might get weird about it.

data Card

A playing card is made up of a rank and a suit (e.g. the 2 of Clubs)

Constructors

Card Rank Suit

rank :: Card -> Rank

suit :: Card -> Suit

type Hand = [Card]

We add a convenience type synonym for a hand of cards.

printHand :: Hand -> String

Render out a list of card names in a longform but readable format

data HandType

Every hand has a "type" which is the highest value hand contained within it.

In the full game of Balatro, there are other hand types, but we'll stick to ones that can be made from a normal deck of cards.

Constructors

None

HighCard

Pair

TwoPair

ThreeOfAKind

Straight

Flush

FullHouse

FourOfAKind

StraightFlush

RoyalFlush

type Chips = Int

type Mult = Int

data PlayOrDiscard

Constructors

Play

Discard

data Move

Constructors

Move PlayOrDiscard [Card]