ProbabilityFunctions

Documentation

factorial :: Int -> Double

Purpose:

Computes n!

Parameters:

1st Parameter (n) : The number top compute the factorial for

Output:

Returns n!

Notes:

n must be greater than zero

nCr :: Int -> Int -> Double

Purpose:

Computes n Choose r (aka the binomial coeficcient)

Parameters:

1st Parameter (n) : The term "n" in the binomial coeficcient

2nd Parameter (r) : The term "r" in the binomial coeficcient

Output:

Returns n Choose r

Notes:

n must be greater than or equal to r

hyperGeometric :: Int -> Int -> Int -> Int -> Double

Purpose:

Computes the probability P(X = w) for the Hypergeometric Probability Distribution X = Hypergeometric(x, y, z); where x is the population size, y is the number of desired elements in the population, z is the number of draws, w is the target number of successes

Parameters:

1st Parameter (x) : The population size

2nd Parameter (y) : The number of desired elements in the population

3rd Parameter (z) : The number of draws / samples taken from the population

4th Parameter (w) : The target number of successes

Output:

Returns P(X = w) for the Hypergeometric Probability Distribution X = Hypergeometric(x, y, z)

Notes:

x must be greater than or equal to y, z and w

y must be greater than or equal to z and w

z must be greater than or equal to w

This function is used to estimate the probabilities of drawing certain hand types. This is because a Hypergerometric Distribution can be used to model the probability of picking w desired elements when we sample without replacement z times from a population of size x with y desired elements in the total population.

See https://en.wikipedia.org/wiki/Hypergeometric_distribution for more information

numAvailableCardsByRank :: Hand -> [Int]

Purpose:

Gets the number of cards of a each rank remaining in the population Parameters:

1st Parameter (playedCards) : A list containing the cards which have already been played

Output:

Returns list containing the number of cards of a each rank remaining in the population

Notes:

playedCards must contain cards which could feasibly occur. For example playedHands could not contain 5 Aces as we know in a deck of cards there are ony 4 cards of each rank

Helper function for probabilityOfHand

probabilityOfHand :: [Card] -> HandType -> Double

Purpose:

Estimates the probability of getting a certain hand type, taking into account which cards have already been used already

Parameters:

1st Parameter (playedCards) : A list containing the cards which have already been played

2nd Parameter : The hand type we are estimating the probability for

Output:

Returns an estimated the probability of getting a certain hand type, taking into account which cards have already been used already

Notes:

playedCards must contain cards which could feasibly occur. For example playedHands could not contain 5 Aces as we know in a deck of cards there are ony 4 cards of each rank

Helper function for probabilityOfImprovement

See "Estimating Probabilities" in the "Halatro AI Approach" section found in index.html

probabilityOfImprovement :: [Card] -> HandType -> Double

Purpose:

Estimates the probability of getting a better hand than the hand that you currently have

Parameters:

1st Parameter (playedCards) : A list containing the cards which have already been played

2nd Parameter : The hand type we are estimating the probability for

Output:

Returns an estimated probability of getting a better hand than the hand that you currently have

Notes:

Helper function for probabilityBasedAI